now with real messaging

+++ HouseService/Models/SyncModels.cs
@@ -12,14 +12,14 @@ namespace HouseService.Models
     {
         public int HouseId { get; set; }
         public string Name { get; set; } = string.Empty;
-        public double Area { get; set; }
+        public decimal Area { get; set; }
     }
 
     public class RoomDto
     {
         public int RoomId { get; set; }
         public int HouseId { get; set; }
-        public double Area { get; set; }
+        public decimal Area { get; set; }
         public string Placement { get; set; } = string.Empty;
     }
 }
